Make tests pass again. Including on mac
Changes the SharedArray to use a normal file instead of a shm/shared
memory file, since it works cross platform.

To maintain the old behaviour, we could prefix the generated file names
with shm:// on linux. Unfortunately this name generation is duplicated
in a few places, so I'm leaving as is for now.